requirements.txt
torch
torchvision
tensorboard
pyyaml
yacs
einops
SimpleITK
monai
mmcv
timm
pandas
numpy
scikit-image
fvcore
